Cleaned up changes
